+++ brest/io/__init__.py
@@ -1,8 +1,10 @@
 from .ios import IO
+from .manio import Manio
 from .usb_relay import USBRelay
 
 __all__ = [
     'IO',
+    'Manio',
     'USBRelay',
 ]